+++ integrations/sync/client/flows/unit-status.ts
@@ -1,7 +1,3 @@
-/**
- * unit status flow
- */
-
 import { StatusValue } from "@snailycad/types";
 import { ClientEvents, SnCommands } from "~/types/events";