+++ integrations/sync/server/server.ts
@@ -1,3 +1,10 @@
+import { getPlayerIds } from "~/utils/get-player-ids.server";
+import "./flows/auth";
+import "./flows/unit-status";
+import "./flows/panic-button";
+import "./flows/911-call-attach";
+import "./flows/traffic-stop";
+
 export function prependSnailyCAD(text: string) {
   return `^8^*[SnailyCAD]:^7^r ${text}`;
 }
@@ -9,10 +16,3 @@ export function getPlayerApiToken(source: number) {
 
   return apiToken;
 }
-
-import { getPlayerIds } from "~/utils/get-player-ids.server";
-import "./flows/auth";
-import "./flows/unit-status";
-import "./flows/panic-button";
-import "./flows/911-call-attach";
-import "./flows/traffic-stop";