+++ src/utils/version.server.ts
@@ -1,4 +1,5 @@
 import { cadRequest } from "./fetch.server";
+import { request } from "undici";
 
 on("onResourceStart", async (resourceName: string) => {
   if (GetCurrentResourceName() !== resourceName) {
@@ -6,7 +7,37 @@ on("onResourceStart", async (resourceName: string) => {
   }
 
   await verifyCADApiVersion();
+  await verifyCurrentVersion();
 });
+
+async function verifyCurrentVersion() {
+  // @ts-expect-error index is not required
+  const currentResourceVersion = GetResourceMetadata(GetCurrentResourceName(), "version");
+
+  console.log(`Checking for updates...
+Current Version: ${currentResourceVersion}`);
+
+  try {
+    const PACKAGE_JSON_URL =
+      "https://raw.githubusercontent.com/SnailyCAD/fivem-integrations/main/package.json";
+
+    const data = (await request(PACKAGE_JSON_URL).then((res) => res.body.json())) as {
+      version: string;
+    };
+
+    if (data.version !== currentResourceVersion) {
+      console.log(`
+---------------------------------------
+
+[${GetCurrentResourceName()}] A new version is available: ${data.version}.
+Please find the latest version at: https://github.com/SnailyCAD/fivem-integrations/releases
+
+---------------------------------------`);
+    }
+  } catch (err) {
+    // ignore
+  }
+}
 
 async function verifyCADApiVersion() {
   try {
@@ -19,7 +50,13 @@ async function verifyCADApiVersion() {
     if (!data?.includes("200 Success. Current CAD Version")) {
       console.warn(`
 ---------------------------------------
-[${GetCurrentResourceName()}] Failed to verify SnailyCAD version. Please make sure you have correctly configured the SnailyCAD URL and API key in your server.cfg. Read more here: https://docs.snailycad.org/docs/fivem-integrations/scripts
+
+[${GetCurrentResourceName()}] Failed to verify SnailyCAD version. Please make sure the following:
+
+- Correct global API Token set in server.cfg as "snaily_cad_api_token"
+- Correct SnailyCAD URL set in server.cfg as "snaily_cad_url"
+- Your SnailyCAD instance API is running and is accessible from this server
+
 ---------------------------------------`);
     }
   } catch (err) {
@@ -27,7 +64,13 @@ async function verifyCADApiVersion() {
 
     console.warn(`
 ---------------------------------------
-[${GetCurrentResourceName()}] Failed to verify SnailyCAD version. Please make sure you have correctly configured the SnailyCAD URL and API key in your server.cfg. Read more here: https://docs.snailycad.org/docs/fivem-integrations/scripts
+
+    [${GetCurrentResourceName()}] Failed to verify SnailyCAD version. Please make sure the following:
+
+    - Correct global API Token set in server.cfg as "snaily_cad_api_token"
+    - Correct SnailyCAD URL set in server.cfg as "snaily_cad_url"
+    - Your SnailyCAD instance API is running and is accessible from this server
+
 ---------------------------------------`);
   }
 }